# BUG: Avoid expensive access of basis.col_status in _highs_wrapper column status loop