@dschmitz89 2025-03-09
Should we mark this as a back port candidate in case there will be 1.15.3?? Fixes a very serious performance regression.